Use as_f64 and as_i64 instead of as_float and as_int
Use as_f64 and as_i64 conversions uniformly for consistency and clarity.
Update EPSILON to f32 and add layout scaling condition to improve node
layout updates.

+++ common/src/constants.rs
@@ -1,1 +1,1 @@
-pub const EPSILON: f64 = 1e-6;
+pub const EPSILON: f32 = 1e-6;
